feat: Add functions to retrieve age and mileage of a car
The code changes add two new functions to retrieve the age and mileage of a car from the `cars_db` database. These functions will be useful for retrieving specific information about cars in the database.

Note: The commit message follows the established convention of starting with a type (feat for feature) and providing a concise and descriptive summary of the changes.